requirements.txt
asgiref==3.2.3
Django==3.0
djangorestframework==3.10.3
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
cykooz.testing=1.0.1
